HACK: turn off UNOPTFLAT lint errors related to execute_valid from the fourstage execute_ctrl unit
This is related to a path in the ORPSoC test system which goes out through the instruction fetch unit, through the RAM slave and back into the LSU which goes straight to the execute_valid signal in the execute stage. This is obviously not a real path as the RAM will never be acking both the instruction and data ports at the same time, but it makes Verilator compile break.

Not sure what the real solution is - either better decoupling from the fetch or data ports of the execute_valid signal.