pipeline ibus_req
in order to utilize the instruction cache efficiently
the upcoming address needs to be known one clock cycle
before the ack, and to be able to ack on every clock cycle
ibus_req needs to be pipelined.

In other words an additional "address" pipeline stage
is added to the instruction fetch unit.