cappuccino: propagate lsu exceptions without asserting lsu_valid
This remedies the situation where a lsu operation is
followed by a jump and the lsu_valid would cause execute_valid
to be asserted, letting the jump assert ctrl_branch_occur
before the lsu exception has time to assert it.
This situation was found in or1k-dsx, causing that test
to fail.