cappuccino: Seperate mem and wb stages from execute stage
The purpose of this is to be able to achieve greater fmax,
and to be able to perform one cycle per instructions
for load/stores. The fmax goal has already been achieved,
but some work remains gaining one-cycle load/stores working.

What those changes does, is to move the load/stores and
mfspr/mtspr ops into mem stage (we still call it ctrl though).
Ideally the data cache should be loaded with the lsu address
directly from the alu adder, but that is still creating a
rather long path (due to that the bypass logic in the cache
is purely combinatorial), so for now the lsu address used
is the registered address, causing one extra stall cycle for
load/stores.