icache, dcache: harmonise icache and dcache
The icache and dcache have somewhat diverted over time,
and their functionality have been a bit different.
That have however changed and they are now functionally
very similiar, but signal names and ways of implementing
that functionality are still different.
This tries to bring them together to ease future maintainance.