cappuccino: resolve branches in decode stage
This moves branch resolving from execute stage into
decode stage and the motivation behind this is to
remove the 1-cycle latency doing it in execute stage
introduces.

To be able to resolve jump to register instructions,
the register port B has to be read in decode stage instead
of execute stage.
When doing that, additional forwarding and locking logic is
necessary and has been implemented accordingly.

coremark and dhrystone benchmarks show around 6-10%
performance increase with those changes.
There is however one downside, the flag generation now
creates a critical path through the ALU back to the branch
indication signal, so some further work is needed to
make the comparison operations more lightweight, or
as a last resort, execute l.sfxx instructions in decode stage.