cappuccino/fetch: always flush on incoming exception
The pipeline_flush_i signal comes one cycle before the
exception branch indication signal on exceptions, but
on the same cycle on l.rfe.
However, the current instruction in fetch will never be
valid on the edge of the exception branch signal, so
we need to push out a nop to decode on this condition.