cappuccino/fetch: remove advance during stall feature
Advancing during stall was added as a performance hack,
to avoid having to refetch a fetch that got stalled
during a bus access, without registering the result.
Currently, we are anyways registering the result, thus
rendering this hack unnecessary.
Besides that, the hack creates a lot of cumbersome
corner cases (i.e. fetch is unstallable under certain
circumstances).