cappuccino/fetch: remove old mmu exception workarounds
When the fetcher was advancing during stall, those were
needed to push out noops when mmu exceptions happened.
This is not necessary anymore and also causes a bug to happen
when a jump is causing an itlb miss and there is a store
currently in ctrl stage that (try to) stall fetch.
What happens then is that the delay slot might not be executed
at all, and the branch destination is considered as a
delay slot instruction, so the exception returns to
branch destination - 4.