cappuccino/ctrl: remove flag write to esr
The situation the comment describes isn't valid anymore,
EPCR will never point past the l.sf but only at it,
so the work-around can be removed.
Not only is the situation not valid, the workaround will
create a bug in case of an exception happening in a delay
slot of a l.bf and the delay slot instruction is an l.sf.
The l.bf will then be rerun with the flag evaluated in the
delay slot.