Rename RAM modules
mor1kx_spram is actually a dual port ram, as it has seperate read and
write ports.
A common notion for such a dual port RAM is "simple dual port ram",
in contrast to the type of dual port ram that has two seperate read
and write ports that's commonly referred to as a "true dual port ram".
This updates the RAM modules to follow this notion.