execute_alu: simplify alu_result_o generation
Instead of decoding opc_insn, just use the control signals for
non adder results to mux alu_result.
This changes the meaning of op_alu from including any operation that
used the ALU to be a select signal for opc_alu, there are anyways no
users left of the old meaning, so this should not cause any troubles.
Also, decoding of movhi op is moved from execute_alu to decode.