cappuccino/decode_execute: insert bubble on l.sfxxx; l.b(n)f
The path from execute that sets the flag to the branch target in address
stage is very long and critical, and not much else than using the
registered flag value is possible to avoid it.
This does that by inserting a bubble on the condition where a
set flag instruction is in execute stage and a conditional branch
instruction is in decode.
Unfortunately, this condition is *very* common, so future steps are
needed to avoid the performance penalty this creates.
Those steps include:
1) Implement simple branch "prediction" by using the old value and
   re-evaluate the branch in execute stage. On mispredictions,
   fetch and decode stage should be flushed and the correct branch
   target should be fetched instead.
2) Make the compilers aware of the problem, so they can insert
   useful instructions in between the set flag and the conditional
   branch instructions.