cappuccino/lsu: use fsm to control external bus accesses
Gives a bit more control over when and how the external bus
accesses are initiated.
Main motivation is for upcoming support for dmmu hw tlb reloads,
where the actual bus access will take place here in the lsu.
It also makes it more in line how things are done in the fetcher,
and enables the possibility to move the cache refill bus access
logic from the cache into the lsu.