icache: reorganize OPTION_ICACHE_WAYS == 2 guard
When OPTION_ICACHE_WAYS == 1, code inside the
OPTION_ICACHE_WAYS == 2 guard contains references to out of
boundary indecies in vectors.
icarus can figure out that this code is not used, if the
OPTION_ICACHE_WAYS == 2 statement is on its own.
There are still problems with verilator, but that needs
a lot more work to fix.