pic: consistent acking for EDGE and LATCHED_LEVEL
Original spec says that '1' is written for clearing interrupt; however
or1200 and most of the software do this the opposite way with '0' for
clearing. In the moment LATCHED_LEVEL conforms to defacto standard
while EDGE works as in spec. Original purpose of these two modes doesn't
reflect the possibility of such discrepancy which may lead to
unnecessary frustration during hacking. This patch changes EDGE trigger
to conform with LATCHED_LEVEL and majority.