dcache: save copy of tag data on refill
The data read from the tag memory might be change during a refill,
so it's necessary to keep a copy around until the data is written back.

9e5a414 Caches: Support multiple ways
removed the old tag saving code, and correctly replaced it in the icache,
but the respective change was omitted from the dcache changes.