cappuccino: add support to read gpr via spr
This adds an extra register file read port, for access to the
gpra via spr.
So far only read access.
The motivation for adding a complete seperate rf_ram is that
it's least intrusive and muxing the address port of one of
the 'real' rf_rams can be avoided.
The downside is of course that it's a bit of a resource waste.