DCache: Coherency
This adds coherency via the snoop port.

The cache have a duplicate tag memory for snooping,
which significantly improve performance over the other
alternative where the normal tag memory is used.

Note by Stefan Kristiansson:
This commit message is an excerpt of the original message
and the commit itself is squashed together from a couple
of trial and errors, but I think it should be
pretty representive of the gist of the original patch.