cappuccino: fix debug unit "asynchronous" stall/resume
Stalling and resuming the CPU at arbitrary places hasn't worked
reliably.
This was especially noticable with the default profiling routines
in openocd, which works by sampling NPC when the cpu is stalled.
To fix this, some more distinctions between stalling when stepping
and stalling/resuming when not stepiing is needed.