icache: move refill ibus logic to fetch
A fair amount of logic in the icache is used to handle
the bus access logic for refills, by moving this logic into the
fetch module (which also need the bus access logic for non-cache
accesses), a bit of resources can be saved.

This only moves the bus access logic, the icache still serves
as a 'master' that initiates the refill.
It might be beneficial to 'dumb-down' the icache even more
and let the fetch module completely control the refill
(and invalidation).