dcache: move refill dbus logic to lsu
As in the similar rework for icache/fetch, the motivation for
this is to save some resources.
An added benefit here is that it *should* also untangle
a pretty critical path going through dc_access out on the dbus
via dbus_adr. With the mux that created that path gone,
hopefully there should be some improvements in timing as well.