pic: make NMIs configurable
That the bottom two irqs are NMIs is something that comes from
the or1200 implementation, there is nowhere stated in the arch
spec about them.
This makes the number of NMIs configurable, to allow backwards
compatibility with or1200, but at the same time, the default
is changed to disable them.