decode: do not generate illegal insn exception on l.msync
l.msync is a class II instruction, which means that it is
optional to implement.
However, this doesn't mean that the instruction should
generate an illegal instruction exception when it's not
implemented, instead the arch spec states that it should
instead be implemented as a no-op on implementations
that has strongly ordered memory accesses.

To be consistent, parameters are added for all three
sync instructions, l.msync, l.psync and l.csync.
None of them have actual implementations, and l.psync and
l.csync will generate illegal insn exceptions when their
parameters are defined as "NONE".