add support for shadow register files
This adds an additional parameter, that configure the number
of shadow register files are present.
This functionality is related to the fast contexts, but
that feature is not yet implemented.

Software can however make use of the shadowed rf without
the fast context functionality, by reading and writing the
shadowed register file by spr accesses.
Reading the RF over spr was already supported,
but this also adds write support.