cappuccino: move reading of rf port A to decode
The benefit of this is that it makes it possible to add an extra
registering stage between decode and execute, avoiding some
critical paths that comes from adding a lot of logic straight to
the rf RAM outputs.

This also allows for the PIPELINED multiplier to hook up
to the outputs at decode stage, to be able to add an extra
registering stage to the end of that.

The disadvantage is that extra forwarding logic is
needed between decode and execute.