cappuccino/lsu: decouple storebuffer from external bus
Instead of exposing the store buffer directly to the external
bus, add an extra registered stage.
This will add a one-cycle latency on the first write when
emptying the store-buffer, but will at the same time
remove critical paths.

This should also make disabling the store buffer all together
more straight forward, but the work to make that possible is not
done here.