cappuccino/lsu: only mask snooped own bus accesses
The assumption that all bus accesses are to snooped devices
(i.e. memory) is not necessary true.
This masks out only the snooped bus accesses that are our own.

In the long run, this masking shouldn't really be in mor1kx,
rather in the snoop unit (outside of mor1kx), and snoop_en_i
should only be asserted when the snoop is valid for this core.