cappuccino/lsu: abort cache refill on snoop hit
Even though the assumption that the bus accesses from the refill
and the snoop are excelusive, there is a possibility that a
refill have been initiated at the same time as the snoop bus
access is going on.
This doesn't cause problems as long as the refill happens
to to the way that is being snooped, but if the snoop occurs to
an opposite way, the refill logic has saved the the old
tag data for that way and will write it back once the refill
has finished.

This avoids the issue by aborting refills when there's
a snoop hit, and prevents refills to be started at the
same instant as the snoop access and the snoop hit is handled.

This can be improved upon in several ways, either by only
aborting/preventing the accesses that are related to the
refill or as an even cleaner solution by splitting the tags
into seperate memories, so the tag saving mechanism can be
removed.

The overhead of aborting all refills on snoop-hits is however
rather small, since the refill would anyway be stalling waiting
for the snooped access to finish, so for now it will suffice as
a solution.