execute_alu: tidy up logic optimization
The optimizations in bba8660
('mor1kx_execute_alu: logic operation optimizations'),
was provided as a proof-of-concept by Florent.
I however decided to pick that as is and do the tidy up in this
seperate commit.
This sprinkles some comments around, fixes a couple of coding
style issues and tries to make it a bit easier to read.