Fix issue where we have an ibus error during cache refill
Found in or1k-tests `or1k-alignillegalinsn`.  When jumping to a bad
address like 0xff000000 the instruction bus raises an error.  At this
point the fetch unit is waiting for the cache fill to be full and never
checks for error causing a lockup.

This change will allow us to detect an error and stop the cache
refill.