Utils: Fix zero as a shadow gpr parameter
For the  mor1kx defaults the shadow GPR's is set to zero, this is used
when calculating the total register file address size but zero is not
supported.  Fix that by supporting 0.

Also, there is an issue with the expand wire assign when there are no
SHADOW SPRs, don't assign in that case.

This fixes github issue #69.