Marocchino: fixup monitoring to have insn inline with data
Previously monitoring was broken because writeback data was not
available when we presented the instrucion to the monitor.  This
meant that the monitor always reported the wrong register values
and result values.

This patch adds some registers to register the insn's and pc's
when then insn is submitted to the order manager, the registers
are then read during writeback.