Fix mulitple monitor block issue with iverilog
I am seeing the following error.

    iverilog -sorpsoc_tb -c mor1kx-generic_1.1.scr -o mor1kx-generic_1.1 -DICARUS_SIM -DSIM -DSPEEDSIM
    ../src/mor1kx_5.0-r3/rtl/verilog/mor1kx_cpu.v:237: error: 'monitor' has already been declared in this scope.
    ../src/mor1kx_5.0-r3/rtl/verilog/mor1kx_cpu.v:204:      : It was declared here as a generate block.

I traced this to us using if conditions where iverilog thinks it's
possible the design will result in multiple monitor blocks.  That
is not the case, but switching to `else if` fixes the issue.

This seems to be related to using a new version of iverilog.