icache: Ensure we reset spr_bus_ack_o on reset
Issue: #127

The reset and error cases override the state to IDLE, but if we
are invalidating the spr_bus_ack_o stays set which should probably
not cause any problem but is showing up as an issue in formal
verification.

Ensure reset and error override invalidate and also reset the ack
output signal on reset.