formal: Update ALU serial op checks to generic bounds checking
The previous verification logic counted the cycles it took
to run a multipy operation stopping when it reached 30. then
confirming we were done.  The new logic is a bit more flexible and
simple we count until the operation is complete.  If the count goes
above 32 we fail.  This can be used for any multi clock
operation so we use for 3 stage as well.

In the .sby file I have updated depth to 64 for serial to allow
k-indection to capture a valid start of a multiply operation.  Also, I
have changed task names to be a bit more descriptive.