Formal: Wishbone Interface
Formally verifying properties of "B3_READ_BURSTING" and "B3_REGISTERED_FEEDBACK" wishbone.
Property checks for IBUS_BRIDGE pass, but DBUS_BRIDGE fails in some cases. "CLASSIC"
Wishbone is not reset, so many properties of this interface are not satisfied. The
maximum request value is set to 16, which can be increased but should choose 'F_LGDEPTH'
carefully, or else the test won't pass f_outstanding validations.