dcache: Fix issue with missing writes during invalidate
This again fixes: #122

Testing to see how it helps with: #146

However, this time I hope it is more stable as formal now caught this
bug and now asserts that the formal properties pass.

The fix is slightly different than the original as now we only change
the state machine to allow going directly from INVALIDATE to WRITE with
write having higher priority over invalidate as per comment.