lsu: Fix error handling and priority pointed out by formal
A few updates to prioritize errors so they correctly reset state and are
not lost.

If we get a dbus_err it's an end game scenario, so we prioritize setting
error codes and moving the LSU state back to IDLE.

Also, this makes refill higher priority than READ.  If we have a bending
refill we want to service is to refresh the cache before starting a
read operation.