ctrl: Update formal to fix on new yosys
This was failing on newer versions of yosys.  The assert is
not true when we are testing the control unit alone as any
input can come in, we update to only check this for CPU level tests.

Also we remove the spr_access_valid assert which is just checking the
properties of an assign.