Formal verification for pfpu32_cmp
This adds formal verification for the comparison unit in the FPU.
Comparison operations don't appear to be fully specified in the ISA
spec, but I took my best stab at sensible semantics, based on my
understanding of IEEE 754.

For the most part, what I came up with lines up with the implementation.
The only difference is the behaviour of `sfult`, `sfule`, `sfugt`, and
`sfuge` when given a signalling NaN. I believe these should result in an
exception (signalled by `inv_o`), but the implementation doesn't do
this.

To get tests to pass, the assertions introduced here have an exception
carved out for this behavioural difference. This exception is disabled
in the `pfpu32_cmp_snan_bug` test, which currently fails.