Verification for pfpu32_muldiv: MULT (and some DIV)
This adds formal verification for the multiplication outputs in
pfpu32_muldiv, and some division outputs too.

To achieve this, I assumed (but not yet verified) that the muldiv module
will not receive further requests for the 12 cycles following a division
request, since the module will be busy with the Goldschmidt algorithm.

To make verification of the multiplication output tractable, I've split
it into two pieces. The `pfpu_muldiv_check` target in the Makefile
checks a series of assertions that together imply an assumption in the
`pfpu_muldiv` target. Unfortunately, making the first set of assertions
isn't enough to guide the solver towards proving the assumption within a
reasonable amount of time.